mbms/requirements.txt
Flask
Flask-API
docker
python-dotenv
PyYAML